again

+++ chia/_tests/core/test_cost_calculation.py
@@ -298,7 +298,7 @@ async def test_get_puzzle_and_solution_for_coin_performance(benchmark_runner: Be
         parent, puzzle, amount_program, _ = spend.as_iter()
         parent_coin_info = parent.as_atom()
         puzzle_hash = puzzle.get_tree_hash()
-        amount = amount_program.as_int()
+        amount = uint64(amount_program.as_int())
         coin = Coin(parent_coin_info=parent_coin_info, puzzle_hash=puzzle_hash, amount=amount)
         spent_coins.append(coin)
