+++ chia/simulator/wallet_tools.py
@@ -148,7 +148,7 @@ class WalletTool:
                     coin_to_append = Coin(
                         coin.name(),
                         bytes32(outputs.vars[0]),
-                        int_from_bytes(outputs.vars[1]),
+                        uint64(int_from_bytes(outputs.vars[1])),
                     )
                     message_list.append(coin_to_append.name())
                 message = std_hash(b"".join(message_list))
